dcc editor: moved dcc widget into its own package
