+++ hswidget/dccwidget/widget.go
@@ -1,4 +1,4 @@
-package hswidget
+package dccwidget
 
 import (
 	"fmt"
