+++ hswindow/hseditor/hsdcceditor/dcc_editor.go
@@ -11,7 +11,7 @@ import (
 	"github.com/OpenDiablo2/HellSpawner/hscommon/hsproject"
 
 	"github.com/OpenDiablo2/HellSpawner/hscommon"
-	"github.com/OpenDiablo2/HellSpawner/hswidget"
+	"github.com/OpenDiablo2/HellSpawner/hswidget/dccwidget"
 
 	"github.com/OpenDiablo2/HellSpawner/hswindow/hseditor"
 )
@@ -45,7 +45,7 @@ func Create(_ *hscommon.TextureLoader,
 // Build builds a dcc editor
 func (e *DCCEditor) Build() {
 	e.IsOpen(&e.Visible).Flags(g.WindowFlagsAlwaysAutoResize).Layout(g.Layout{
-		hswidget.DCCViewer(e.Path.GetUniqueID(), e.dcc),
+		dccwidget.DCCViewer(e.Path.GetUniqueID(), e.dcc),
 	})
 }
